refactor data-mgt implementation services
