+++ src/app/api/data-mgt/local-memory/impl/shared.local-memory-data-api.service.ts
@@ -8,7 +8,7 @@ import { Observable, of } from 'rxjs';
 import { ProductFamily } from 'src/app/models/entities/ProductFamily';
 import { ProductType } from 'src/app/models/entities/ProductType';
 import { SellType } from 'src/app/models/entities/SellType';
-import { SharedDataApiIService } from '../shared-data-api.iservice';
+import { SharedDataApiIService } from '../../shared-data-api.iservice';
 import { Person } from 'src/app/models/entities/Person';
 
 export const MOCK_PEOPLE: Partial<Person>[] = [
