+++ src/app/api/data-mgt/local-memory/local-memory-data-api.module.ts
@@ -5,12 +5,12 @@
 
 import { NgModule } from '@angular/core';
 import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
-import { ClientsLocalMemoryDataApiService } from './entity-impl/clients.local-memory-data-api.service';
-import { SellersLocalMemoryDataApiService } from './entity-impl/sellers.local-memory-data-api.service';
-import { ProductsLocalMemoryDataApiService } from './entity-impl/products.local-memory-data-api.service';
-import { SalesLocalMemoryDataApiService } from './entity-impl/sales.local-memory-data-api.service';
-import { SharedLocalMemoryDataService } from './shared.local-memory-data-api.service';
-import { UsersLocalMemoryDataApiService } from './entity-impl/users.local-memory-data-api.service';
+import { ClientsLocalMemoryDataApiService } from './impl/clients.local-memory-data-api.service';
+import { SellersLocalMemoryDataApiService } from './impl/sellers.local-memory-data-api.service';
+import { ProductsLocalMemoryDataApiService } from './impl/products.local-memory-data-api.service';
+import { SalesLocalMemoryDataApiService } from './impl/sales.local-memory-data-api.service';
+import { SharedLocalMemoryDataService } from './impl/shared.local-memory-data-api.service';
+import { UsersLocalMemoryDataApiService } from './impl/users.local-memory-data-api.service';
 import { LocalMemoryDataAccessApiService } from './local-memory-data-access-api.service';
 
 /**
